move secs_to_str to utils, add test
