+++ hive/utils/timer.py
@@ -1,6 +1,7 @@
 """Timer for reporting progress on long batch operations."""
 
 import time
+from hive.utils.normalize import secs_to_str
 
 class Timer:
     """Times long routines, printing status and ETA.
@@ -73,7 +74,7 @@ class Timer:
         else:
             total_time = self._end_time - self._start_time
             out += "done in %s, avg rate: %.1f/s" % (
-                Timer.secs_to_str(total_time),
+                secs_to_str(total_time),
                 self._total / total_time)
 
         return out
@@ -87,22 +88,7 @@ class Timer:
         """Time to finish, based on most recent batch."""
         left = self._full_total - self._processed
         secs = (left / self._rate())
-        return Timer.secs_to_str(secs)
-
-    @staticmethod
-    def secs_to_str(secs):
-        """Given number of seconds returns, e.g., `02h 29m 39s`"""
-        units = (('s', 60), ('m', 60), ('h', 24), ('d', 7))
-        out = []
-        rem = secs
-        for (unit, cycle) in units:
-            out.append((rem % cycle, unit))
-            rem = int(rem / cycle)
-            if not rem:
-                break
-        if rem: # leftover = weeks
-            out.append((rem, 'w'))
-        return ' '.join(["%02d%s" % tup for tup in out[::-1]])
+        return secs_to_str(secs)
 
     def _elapsed(self, lap_idx=None):
         if not lap_idx:
